page/recordpage: added RecordProgressProvider
